+++ ariadne_codegen/client_generators/package.py
@@ -1,7 +1,7 @@
 import ast
 from datetime import datetime
 from pathlib import Path
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Set
 
 from graphql import FragmentDefinitionNode, GraphQLSchema, OperationDefinitionNode
 
@@ -162,6 +162,8 @@ class PackageGenerator:
         )
         self.scalars_definitions_file_name = "scalars"
 
+        self._unpacked_fragments: Set[str] = set()
+
     def generate(self) -> List[str]:
         """Generate package with graphql client."""
         self._validate_unique_file_names()
@@ -203,6 +205,9 @@ class PackageGenerator:
             convert_to_snake_case=self.convert_to_snake_case,
             custom_scalars=self.custom_scalars,
             plugin_manager=self.plugin_manager,
+        )
+        self._unpacked_fragments = self._unpacked_fragments.union(
+            query_types_generator.get_unpacked_fragments()
         )
         self.result_types_files[file_name] = query_types_generator.generate()
         operation_str = query_types_generator.get_operation_as_str()
@@ -313,13 +318,16 @@ class PackageGenerator:
             self.generated_files.append(file_path.name)
 
     def _generate_fragments(self):
-        if not self.fragments_definitions:
+        if not set(self.fragments_definitions.keys()).difference(
+            self._unpacked_fragments
+        ):
             return
 
         generator = FragmentsGenerator(
             schema=self.schema,
             enums_module_name=self.enums_module_name,
             fragments_definitions=self.fragments_definitions,
+            exclude_names=self._unpacked_fragments,
             base_model_import=self.base_model_import,
             convert_to_snake_case=self.convert_to_snake_case,
             custom_scalars=self.custom_scalars,