+++ tests/main/clients/inline_fragments/expected_client/__init__.py
@@ -8,6 +8,15 @@ from .exceptions import (
     GraphQLClientHttpError,
     GraphQlClientInvalidResponseError,
 )
+from .fragments import (
+    FragmentOnQueryWithInterface,
+    FragmentOnQueryWithInterfaceQueryIInterface,
+    FragmentOnQueryWithInterfaceQueryITypeA,
+    FragmentOnQueryWithInterfaceQueryITypeB,
+    FragmentOnQueryWithUnion,
+    FragmentOnQueryWithUnionQueryUTypeA,
+    FragmentOnQueryWithUnionQueryUTypeB,
+)
 from .interface_a import (
     InterfaceA,
     InterfaceAQueryIInterface,
@@ -16,6 +25,21 @@ from .interface_a import (
 )
 from .interface_b import InterfaceB, InterfaceBQueryIInterface, InterfaceBQueryITypeA
 from .interface_c import InterfaceC, InterfaceCQueryI
+from .query_with_fragment_on_interface import (
+    QueryWithFragmentOnInterface,
+    QueryWithFragmentOnInterfaceQueryIInterface,
+    QueryWithFragmentOnInterfaceQueryITypeA,
+    QueryWithFragmentOnInterfaceQueryITypeB,
+)
+from .query_with_fragment_on_query_with_interface import (
+    QueryWithFragmentOnQueryWithInterface,
+)
+from .query_with_fragment_on_query_with_union import QueryWithFragmentOnQueryWithUnion
+from .query_with_fragment_on_union import (
+    QueryWithFragmentOnUnion,
+    QueryWithFragmentOnUnionQueryUTypeA,
+    QueryWithFragmentOnUnionQueryUTypeB,
+)
 from .union_a import UnionA, UnionAQueryUTypeA, UnionAQueryUTypeB
 from .union_b import UnionB, UnionBQueryUTypeA, UnionBQueryUTypeB
 from .union_c import UnionC, UnionCQueryUTypeA, UnionCQueryUTypeB
@@ -24,6 +48,13 @@ __all__ = [
     "AsyncBaseClient",
     "BaseModel",
     "Client",
+    "FragmentOnQueryWithInterface",
+    "FragmentOnQueryWithInterfaceQueryIInterface",
+    "FragmentOnQueryWithInterfaceQueryITypeA",
+    "FragmentOnQueryWithInterfaceQueryITypeB",
+    "FragmentOnQueryWithUnion",
+    "FragmentOnQueryWithUnionQueryUTypeA",
+    "FragmentOnQueryWithUnionQueryUTypeB",
     "GraphQLClientError",
     "GraphQLClientGraphQLError",
     "GraphQLClientGraphQLMultiError",
@@ -38,6 +69,15 @@ __all__ = [
     "InterfaceBQueryITypeA",
     "InterfaceC",
     "InterfaceCQueryI",
+    "QueryWithFragmentOnInterface",
+    "QueryWithFragmentOnInterfaceQueryIInterface",
+    "QueryWithFragmentOnInterfaceQueryITypeA",
+    "QueryWithFragmentOnInterfaceQueryITypeB",
+    "QueryWithFragmentOnQueryWithInterface",
+    "QueryWithFragmentOnQueryWithUnion",
+    "QueryWithFragmentOnUnion",
+    "QueryWithFragmentOnUnionQueryUTypeA",
+    "QueryWithFragmentOnUnionQueryUTypeB",
     "UnionA",
     "UnionAQueryUTypeA",
     "UnionAQueryUTypeB",