+++ tests/main/clients/inline_fragments/expected_client/query_with_fragment_on_query_with_interface.py
@@ -1,0 +1,8 @@
+from .fragments import FragmentOnQueryWithInterface
+
+
+class QueryWithFragmentOnQueryWithInterface(FragmentOnQueryWithInterface):
+    pass
+
+
+QueryWithFragmentOnQueryWithInterface.update_forward_refs()