+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed `ResultTypesGenerator` to trigger `generate_result_class` for each result model.
 - Changed processing of models fields to trim leading underscores.
 - Added `ShorterResultsPlugin` to standard plugins.
+- Fixed handling of inline fragments inside other fragments.
 
 
 ## 0.6.0 (2023-04-18)